+++ changelog.md
@@ -3,4 +3,4 @@ Ported to 1.21.1
 
 * Formulas for over time strength has been adjusted due to saturation_modifier no longer existing
   * Overall speed with overtime regen reduced
-* Cakes now heal 50% missing health instead of 30%, but over time and no longer instantly
+* Cakes now heal 50% missing health instead of 30% and minimum 2 instead of 1, but over time and no longer instantly